Update the Json conversion methods
The tasks now have a way identify the differences, between recurring, transient and anti-task, when converting from Json to Java objects or vice versa. Also, I added two methods to Task.java, getType() and isRecurringTask(), to check if the task is  a recurring task. And I added three methods to RecurringTask.java, getEndDate(), getFrequency(), and an overriding method of Task.java,  isRecurringTask().